services/loan_portal_bff/requirements.txt
strawberry-graphql[fastapi]
uvicorn[standard]
httpx
aiodataloader
pydantic-settings
